Railways conf/Production/requirements.txt
numpy>=1.20.0
scipy>=1.7.0
scikit-signal>=0.2.0
pyyaml>=5.4
matplotlib>=3.3.0
